dynamic adding and deleting a post

+++ Major Project/passport.js/codeial/controllers/post_controller.js
@@ -4,13 +4,23 @@ const Comment = require("../models/comment");
 
 module.exports.create = async function (req, res) {
   try {
-    await Post.create({
+    let post = await Post.create({
       content: req.body.content,
       user: req.user._id,
     });
+    if (req.xhr) {
+      return res.status(200).json({
+        data: {
+          post: post,
+        },
+        message: "Post Created",
+      });
+    }
+    req.flash("success", "post created");
+
     return res.redirect("back");
   } catch (err) {
-    console.log("Couldnt create a post", err);
+    req.flash("error", "Cannot be posted");
     return res.redirect("/");
   }
 };
@@ -21,13 +31,29 @@ module.exports.destroy = async function (req, res) {
 
     if (req.user.id == post.user) {
       post.remove();
+
+      if (req.xhr) {
+        return res.status(200).json({
+          data: {
+            post_id: req.params.id,
+          },
+          message: "Post deleted",
+        });
+      }
+
+      req.flash("success", "post deleted");
+
       await Comment.deleteMany(req.params.id);
+
       return res.redirect("back");
     } else {
+      req.flash("error", "You cannot delte the post");
+
       return res.redirect("back");
     }
   } catch (err) {
-    console.log(err);
+    req.flash("error", err);
+
     return res.redirect("back");
   }
 };
